# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plurality API",
    "description": "Description of Plurality API",
    "version": "1.0.0"
  },
  "host": "app.plurality.local:443",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/user/auth/otp/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/user/auth/otp/authenticate": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/oauth-twitter/": {
      "get": {
        "tags": [
          "OAuth-Twitter"
        ],
        "description": "",
        "parameters": [
          {
            "name": "sse_id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/oauth-twitter/event": {
      "post": {
        "tags": [
          "OAuth-Twitter"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-sse-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth-twitter/info": {
      "get": {
        "tags": [
          "OAuth-Twitter"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-token-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "408": {
            "description": "Request Timeout"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}